+++ test/image_tags_test.js
@@ -171,6 +171,181 @@ describe("ImageTags", function() {
     );
   });
 
+  it("gets story hero image attribution values instead of card attribution values on story share", function() {
+    const story = {
+      "hero-image-s3-key": "my/image.png",
+      "hero-image-attribution": "attribution test",
+      cards: [
+        {
+          id: "sample-card-id",
+          metadata: {
+            "social-share": {
+              title: "share-card-title",
+              message: "share-card-description",
+              image: {
+                key: "my/card/image.jpg",
+                attribution: null,
+                metadata: {
+                  width: 1300,
+                  height: 1065,
+                  "mime-type": "image/jpeg"
+                }
+              }
+            }
+          }
+        }
+      ]
+    };
+
+    const opts = {};
+
+    const string = getSeoMetadata(
+      seoConfig,
+      config,
+      "story-page",
+      { data: { story: story } },
+      opts
+    );
+
+    assertContains('<meta property="twitter:image:alt" content="attribution test"/>', string)
+
+    assertContains('<meta property="og:image:alt" content="attribution test"/>', string)
+
+  });
+
+  it("gets story summary as attribution if hero image attribution is not present", function() {
+    const story = {
+      "hero-image-s3-key": "my/image.png",
+      "hero-image-attribution": null,
+      summary: "attribution test",
+      cards: [
+        {
+          id: "sample-card-id",
+          metadata: {
+            "social-share": {
+              title: "share-card-title",
+              message: "share-card-description",
+              image: {
+                key: "my/card/image.jpg",
+                metadata: {
+                  width: 1300,
+                  height: 1065,
+                  "mime-type": "image/jpeg"
+                }
+              }
+            }
+          }
+        }
+      ]
+    };
+
+    const opts = {};
+
+    const string = getSeoMetadata(
+      seoConfig,
+      config,
+      "story-page",
+      { data: { story: story } },
+      opts
+    );
+
+    assertContains('<meta property="twitter:image:alt" content="attribution test"/>', string)
+
+    assertContains('<meta property="og:image:alt" content="attribution test"/>', string)
+
+  });
+
+  it("gets story headline as attribution if hero image attribution and summary is not present", function() {
+    const story = {
+      "hero-image-s3-key": "my/image.png",
+      "hero-image-attribution": null,
+      summary: null,
+      headline: "attribution test",
+      cards: [
+        {
+          id: "sample-card-id",
+          metadata: {
+            "social-share": {
+              title: "share-card-title",
+              message: "share-card-description",
+              image: {
+                key: "my/card/image.jpg",
+                metadata: {
+                  width: 1300,
+                  height: 1065,
+                  "mime-type": "image/jpeg"
+                }
+              }
+            }
+          }
+        }
+      ]
+    };
+
+    const opts = {};
+
+    const string = getSeoMetadata(
+      seoConfig,
+      config,
+      "story-page",
+      { data: { story: story } },
+      opts
+    );
+
+    assertContains('<meta property="twitter:image:alt" content="attribution test"/>', string)
+
+    assertContains('<meta property="og:image:alt" content="attribution test"/>', string)
+
+  });
+
+  it("gets card image attribution values instead of story attribution values on card share", function() {
+    const story = {
+      "hero-image-s3-key": "my/image.png",
+      headline: "story headline",
+      cards: [
+        {
+          id: "sample-card-id",
+          metadata: {
+            "social-share": {
+              title: "share-card-title",
+              message: "share-card-description",
+              image: {
+                attribution: "attribution test",
+                key: "my/card/image.jpg",
+                metadata: {
+                  width: 1300,
+                  height: 1065,
+                  "mime-type": "image/jpeg"
+                }
+              }
+            }
+          }
+        }
+      ]
+    };
+
+    const opts = {
+      url: {
+        query: {
+          cardId: "sample-card-id"
+        }
+      }
+    };
+
+    const string = getSeoMetadata(
+      seoConfig,
+      config,
+      "story-page",
+      { data: { story: story } },
+      opts
+    );
+
+    assertContains('<meta property="twitter:image:alt" content="attribution test"/>', string)
+
+    assertContains('<meta property="og:image:alt" content="attribution test"/>', string)
+
+  });
+
   it("gets story data as fallback if the card metadata is falsy", function() {
     const story = {
       "hero-image-s3-key": "my/image.png",
@@ -251,6 +426,38 @@ describe("ImageTags", function() {
       );
       assertContains(
         '<meta property="og:image:height" content="630"/>',
+        string
+      );
+    });
+
+    it("pulls images with alt from the collection if present", function() {
+      const collection = {
+        name: "collection name",
+        summary: "collection summary",
+        metadata: {
+          "cover-image": {
+            "cover-image-s3-key": "my/image.png",
+            "cover-image-metadata": {
+              width: 2400,
+              height: 1260,
+              "focus-point": [0, 0]
+            }
+          }
+        }
+      };
+      const string = getSeoMetadata(
+        seoConfig,
+        config,
+        "home-page",
+        { data: { collection: collection } },
+        {}
+      );
+      assertContains(
+        '<meta property="og:image:alt" content="collection summary"/>',
+        string
+      );
+      assertContains(
+        '<meta property="twitter:image:alt" content="collection summary"/>',
         string
       );
     });