+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.38.40](https://github.com/quintype/quintype-node-seo/compare/v1.38.39...v1.38.40) (2021-07-28)
+
+
+### Features
+
+* **Structure data tool testing:** Add support for testing structure data with all necessary schemas ⚡  ([#507](https://github.com/quintype/quintype-node-seo/issues/507)) ([aeea32a](https://github.com/quintype/quintype-node-seo/commit/aeea32a994bf4cf64cdf538d2ec3d785203b3817))
+
+### [1.38.39](https://github.com/quintype/quintype-node-seo/compare/v1.38.38...v1.38.39) (2021-07-27)
+
+
+### Bug Fixes
+
+* **pickImage:** image is undefined of pickImage ([#508](https://github.com/quintype/quintype-node-seo/issues/508)) ([fecc8be](https://github.com/quintype/quintype-node-seo/commit/fecc8be44c86220456a511a89d197b52d60875c6))
+
+### [1.38.38](https://github.com/quintype/quintype-node-seo/compare/v1.38.37...v1.38.38) (2021-07-26)
+
+
+### Features
+
+* **ImageTags:** Added support for og:image:alt and twitter:image:alt ([#501](https://github.com/quintype/quintype-node-seo/issues/501)) ([3d271ab](https://github.com/quintype/quintype-node-seo/commit/3d271ab21e009e0520ce1e6949bb1c8d32eeefa1))
+
 ### [1.38.37](https://github.com/quintype/quintype-node-seo/compare/v1.38.37-fix-canonical-url.0...v1.38.37) (2021-07-15)
 
 ### [1.38.36](https://github.com/quintype/quintype-node-seo/compare/v1.38.35...v1.38.36) (2021-06-17)